requirements.txt
Requests
apscheduler
pyyaml
mypy
types-PyYAML
types-requests
loguru